private folder로 directory 정리 및 공통 component화 하기
